fmt and clippy
