+++ build.rs
@@ -60,13 +60,12 @@ const PIECEWISE_CUTOFFS: [f64; 15] = [
 /// Section: "Initial approximation for the F-interval [0, 5)"
 /// Equation 4
 fn generate_sinh_approximator() -> String {
-    let mut code = format!(
-        r"
+    let mut code = r"
         // Generated by build.rs:generate_sinh_approximator
         #![allow(clippy::all)]
         #![allow(dead_code)]
 
-        pub struct SinhApprParams {{
+        pub struct SinhApprParams {
             pub a: f64,
             pub s: f64,
             pub p_1: f64,
@@ -74,11 +73,11 @@ fn generate_sinh_approximator() -> String {
             pub p_3: f64,
             pub q_1: f64,
             pub q_2: f64,
-        }}
+        }
 
         pub fn sinh_approx_lt5_inner(
             f: f64, args: SinhApprParams
-        ) -> f64 {{
+        ) -> f64 {
             let f_min_a = f - args.a;
             let f_min_a_sq = f_min_a * f_min_a;
             let f_min_a_cu = f_min_a * f_min_a_sq;
@@ -95,17 +94,17 @@ fn generate_sinh_approximator() -> String {
                 args.q_2 * f_min_a_sq;
 
             return numerator / denominator;
-        }}    
+        }
 
         /// Approximates the sinh(F) function for F in the interval [0, 5).
-        pub fn sinh_approx_lt5(f: f64) -> f64 {{
+        pub fn sinh_approx_lt5(f: f64) -> f64 {
             sinh_approx_lt5_inner(f, get_sinh_approx_params(f))
-        }}
+        }
 
         /// Returns the parameters for the sinh approximation for the given point.
-        pub fn get_sinh_approx_params(f: f64) -> SinhApprParams {{
-        ",
-    );
+        pub fn get_sinh_approx_params(f: f64) -> SinhApprParams {
+        "
+    .to_string();
 
     for i in 0..PIECEWISE_VALUES.len() {
         // sinh(F) ~= P_a(F) ===
@@ -152,5 +151,5 @@ fn generate_sinh_approximator() -> String {
     code += "{ unreachable!(); } }";
     code += format!("pub const SINH_5: f64 = {};", 5.0_f64.sinh()).as_str();
 
-    return code;
+    code
 }
